Rename useActiveCategory hook
