+++ specifyweb/frontend/js_src/lib/components/UserPreferences/Aside.tsx
@@ -1,83 +1,13 @@
 import React from 'react';
 import { useLocation } from 'react-router';
 import { useNavigate } from 'react-router-dom';
-import _ from 'underscore';
 
 import { listen } from '../../utils/events';
-import type { GetSet, WritableArray } from '../../utils/types';
+import type { GetSet } from '../../utils/types';
 import { Link } from '../Atoms/Link';
 import { locationToState } from '../Router/RouterState';
 import { usePrefDefinitions } from './index';
 
-/**
- * Update the active category on the sidebar as user scrolls
- *
- * Previous implementation used IntersectionObserver for this, but it was too
- * slow when you have 180 categories (on the Data Model page)
- */
-export function useActiveCategory(): {
-  readonly visibleChild: number | undefined;
-  readonly forwardRefs: (index: number, element: HTMLElement | null) => void;
-  readonly scrollContainerRef: React.RefCallback<HTMLDivElement | null>;
-} {
-  const [activeCategory, setActiveCategory] = React.useState<
-    number | undefined
-  >(0);
-  const references = React.useRef<WritableArray<HTMLElement | undefined>>([]);
-
-  const [container, setContainer] = React.useState<HTMLDivElement | null>(null);
-  React.useEffect(() => {
-    if (container === null) return undefined;
-
-    function rawHandleChange(): void {
-      if (container === null) return;
-      const { x, y } = container.getBoundingClientRect();
-      const visibleElement = document.elementFromPoint(
-        x,
-        y + marginTop
-      ) as HTMLElement;
-      if (visibleElement === null) return;
-      const section = findSection(container, visibleElement);
-      if (section === undefined) return;
-      const index = references.current.indexOf(section);
-      setActiveCategory(index === -1 ? undefined : index);
-    }
-
-    const handleChange = _.throttle(rawHandleChange, scrollThrottle);
-
-    const observer = new ResizeObserver(handleChange);
-    observer.observe(container);
-    const scroll = listen(container, 'scroll', handleChange);
-    return (): void => {
-      observer.disconnect();
-      scroll();
-    };
-  }, [container]);
-
-  return {
-    visibleChild: activeCategory,
-    forwardRefs: React.useCallback((index, element) => {
-      references.current[index] = element ?? undefined;
-    }, []),
-    scrollContainerRef: setContainer,
-  };
-}
-
-/**
- * Look for an element this many pixels below the top of the scroll container
- */
-const marginTop = 200;
-const scrollThrottle = 50;
-
-function findSection(
-  container: HTMLElement,
-  child: HTMLElement
-): HTMLElement | undefined {
-  const parent = child.parentElement;
-  if (parent === container) return child;
-  else if (parent === null) return undefined;
-  else return findSection(container, parent);
-}
 export function PreferencesAside({
   activeCategory,
 }: {
